requirements.txt
langchain_core
langchain_openai
streamlit
python-dotenv
langchain_community
langserve
sse_starlette
uvicorn
langchain